consolate esp32 to esp folder, #define GPIO or RMT driver, smaller crc8 table
