+++ examples/drivers/onewire/manifest.json
@@ -1,7 +1,6 @@
 {
 	"include": [
 		"$(MODDABLE)/examples/manifest_base.json",
-		"$(MODULES)/pins/digital/manifest.json"			
 	],
 	"modules": {
 		"*": [
@@ -14,18 +13,10 @@
 		"onewire",
 		"DS18X20"
 	],
-	"defines": {
-		"onewire": {
-			"rmt_rx_channel": "RMT_CHANNEL_2",
-			"rmt_tx_channel": "RMT_CHANNEL_3",
-		},
-	},
 	"platforms": {
 		"esp": {
 			"include": [
-                "$(MODDABLE)/examples/manifest_base.json",
-				"$(MODDABLE)/examples/manifest_net.json",
-				"$(MODULES)/pins/digital/manifest.json"				
+				"$(MODULES)/pins/digital/manifest.json"
 			],
 			"modules": {
 				"*": [
@@ -35,22 +26,37 @@
 			"config": {
 				"onewire": {
 					"pin": "4"
+				},
+			},
+			"defines": {
+				"onewire": {
+					"driver_gpio": "GPIO",
 				}
 			}
 		},
 		"esp32": {
+			"include": [
+				"$(MODULES)/pins/digital/manifest.json"
+			],
 			"modules": {
-				"*": "$(MODULES)/drivers/onewire/esp32/*",
+				"*": "$(MODULES)/drivers/onewire/esp/*",
 			},
 			"config": {
 				"onewire": {
 					"pin": "16"
-				}
-			}
+				},
+			},
+			"defines": {
+				"onewire": {
+					"driver_rmt": "RMT",
+					"rmt_rx_channel": "RMT_CHANNEL_2",
+					"rmt_tx_channel": "RMT_CHANNEL_3",
+				},
+			},
 		},
 		"win": {
 			"modules": {
-				"*": "$(MODULES)/drivers/onewire/win/*",
+				"*": "$(MODULES)/drivers/onewire/sim/*",
 			},
 			"config": {
 				"onewire": {
