+++ modules/drivers/onewire/esp/owb.c
@@ -3,6 +3,7 @@
  *
  * Copyright (c) 2017 David Antliff
  * Copyright (c) 2017 Chris Morgan <chmorgan@gmail.com>
+ * Copyright (c) 2019/05/11  Wilberforce for use in Moddable SDK
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -35,15 +36,6 @@
 
 #include "modGPIO.h"
 
-
-/*
-root@office-pc:/mnt/c/Users/rhys/Projects/moddable/build/tmp/esp/debug/onewire# addr2line -e main.elf 0x402303d3 3ffee090:  3fff5abc 3fff803c 40262ebd 4021d1ac
-/cygdrive/c/Users/rhys/Projects/moddable/examples/drivers/onewire/C:\Users\rhys\Projects\moddable\build\devices\esp\lib\tinyprintf/tinyprintf.c:500
-500:   data->dest[data->num_chars++] = c;
-*/
-
-//#include "tinyprintf.h"
-
 #include "owb.h"
 #include "owb_gpio.h"
 
@@ -74,41 +66,35 @@ static bool _is_init(const OneWireBus * bus)
  * @param[in] crc Starting CRC value. Pass in prior CRC to accumulate.
  * @param[in] data Byte to feed into CRC.
  * @return Resultant CRC value.
+ Dow-CRC using polynomial X^8 + X^5 + X^4 + X^0
+ Tiny 2x16 entry CRC table created by Arjen Lentz
+ See http://lentz.com.au/blog/calculating-crc-with-a-tiny-32-entry-lookup-table
+ */
+static const uint8_t dscrc2x16_table[] = {
+	0x00, 0x5E, 0xBC, 0xE2, 0x61, 0x3F, 0xDD, 0x83,
+	0xC2, 0x9C, 0x7E, 0x20, 0xA3, 0xFD, 0x1F, 0x41,
+	0x00, 0x9D, 0x23, 0xBE, 0x46, 0xDB, 0x65, 0xF8,
+	0x8C, 0x11, 0xAF, 0x32, 0xCA, 0x57, 0xE9, 0x74
+};
+
+static uint8_t _calc_crc_block(uint8_t crc, const uint8_t *buffer, size_t len)
+{
+	while (len--) {
+		crc = *buffer++ ^ crc;  // just re-using crc as intermediate
+		crc = (dscrc2x16_table[crc & 0x0f]) ^ (dscrc2x16_table[16 + ((crc >> 4) & 0x0f)]);
+	}
+  return crc;
+}
+
+/**
+ * @brief 1-Wire 8-bit CRC lookup.
+ * @param[in] crc Starting CRC value. Pass in prior CRC to accumulate.
+ * @param[in] data Byte to feed into CRC.
+ * @return Resultant CRC value.
  */
 static uint8_t _calc_crc(uint8_t crc, uint8_t data)
 {
-    // https://www.maximintegrated.com/en/app-notes/index.mvp/id/27
-    static const uint8_t table[256] = {
-            0, 94, 188, 226, 97, 63, 221, 131, 194, 156, 126, 32, 163, 253, 31, 65,
-            157, 195, 33, 127, 252, 162, 64, 30, 95, 1, 227, 189, 62, 96, 130, 220,
-            35, 125, 159, 193, 66, 28, 254, 160, 225, 191, 93, 3, 128, 222, 60, 98,
-            190, 224, 2, 92, 223, 129, 99, 61, 124, 34, 192, 158, 29, 67, 161, 255,
-            70, 24, 250, 164, 39, 121, 155, 197, 132, 218, 56, 102, 229, 187, 89, 7,
-            219, 133, 103, 57, 186, 228, 6, 88, 25, 71, 165, 251, 120, 38, 196, 154,
-            101, 59, 217, 135, 4, 90, 184, 230, 167, 249, 27, 69, 198, 152, 122, 36,
-            248, 166, 68, 26, 153, 199, 37, 123, 58, 100, 134, 216, 91, 5, 231, 185,
-            140, 210, 48, 110, 237, 179, 81, 15, 78, 16, 242, 172, 47, 113, 147, 205,
-            17, 79, 173, 243, 112, 46, 204, 146, 211, 141, 111, 49, 178, 236, 14, 80,
-            175, 241, 19, 77, 206, 144, 114, 44, 109, 51, 209, 143, 12, 82, 176, 238,
-            50, 108, 142, 208, 83, 13, 239, 177, 240, 174, 76, 18, 145, 207, 45, 115,
-            202, 148, 118, 40, 171, 245, 23, 73, 8, 86, 180, 234, 105, 55, 213, 139,
-            87, 9, 235, 181, 54, 104, 138, 212, 149, 203, 41, 119, 244, 170, 72, 22,
-            233, 183, 85, 11, 136, 214, 52, 106, 43, 117, 151, 201, 74, 20, 246, 168,
-            116, 42, 200, 150, 21, 75, 169, 247, 182, 232, 10, 84, 215, 137, 107, 53
-    };
-
-    return table[crc ^ data];
-}
-
-static uint8_t _calc_crc_block(uint8_t crc, const uint8_t * buffer, size_t len)
-{
-    do
-    {
-        crc = _calc_crc(crc, *buffer++);
-        // ("crc 0x%02x, len %d", (int)crc, (int)len);
-    }
-    while (--len > 0);
-    return crc;
+    return crc = (dscrc2x16_table[crc & 0x0f]) ^ (dscrc2x16_table[16 + ((crc >> 4) & 0x0f)]);
 }
 
 /**
@@ -277,7 +263,6 @@ owb_status owb_use_crc(OneWireBus * bus, bool use_crc)
     } else
     {
         bus->use_crc = use_crc;
-        // ("use_crc %d", bus->use_crc);
 
         status = OWB_STATUS_OK;
     }
@@ -311,7 +296,6 @@ owb_status owb_read_rom(const OneWireBus * bus, OneWireBus_ROMCode *rom_code)
             {
                 if (owb_crc8_bytes(0, rom_code->bytes, sizeof(OneWireBus_ROMCode)) != 0)
                 {
-                    //ESP_LOGE(TAG, "CRC failed");
                     memset(rom_code->bytes, 0, sizeof(OneWireBus_ROMCode));
                     status = OWB_STATUS_CRC_FAILED;
                 } else
@@ -322,14 +306,10 @@ owb_status owb_read_rom(const OneWireBus * bus, OneWireBus_ROMCode *rom_code)
             {
                 status = OWB_STATUS_OK;
             }
-            //char rom_code_s[OWB_ROM_CODE_STRING_LENGTH];
-            //owb_string_from_rom_code(*rom_code, rom_code_s, sizeof(rom_code_s));
-            // ("rom_code %s", rom_code_s);
         }
         else
         {
             status = OWB_STATUS_DEVICE_NOT_RESPONDING;
-            //(TAG, "ds18b20 device not responding");
         }
     }
 
@@ -560,14 +540,4 @@ owb_status owb_search_next(const OneWireBus * bus, OneWireBus_SearchState * stat
     }
 
     return status;
-}
-
-char * owb_string_from_rom_code(OneWireBus_ROMCode rom_code, char * buffer, size_t len)
-{
-    for (int i = sizeof(rom_code.bytes) - 1; i >= 0; i--)
-    {
-        sprintf(buffer, "%02x", rom_code.bytes[i]);
-        buffer += 2;
-    }
-    return buffer;
 }
